requirements.txt
lief
gitpython